+++ server/src/main/java/org/opensearch/cluster/metadata/MetadataUpdateSettingsService.java
@@ -304,7 +304,10 @@ public class MetadataUpdateSettingsService {
                             for (Index index : request.indices()) {
                                 if (index.getName().charAt(0) != '.') {
                                     // No replica count validation for system indices
-                                    Optional<String> error = awarenessReplicaBalance.validate(updatedNumberOfSearchReplicas);
+                                    Optional<String> error = awarenessReplicaBalance.validate(
+                                        updatedNumberOfSearchReplicas,
+                                        AutoExpandSearchReplicas.SETTING.get(openSettings)
+                                    );
 
                                     if (error.isPresent()) {
                                         ValidationException ex = new ValidationException();